DataAggregateInfra/web/requirements.txt
flask
psycopg2-binary
plotly
numpy
pandas
datetime